+++ lib/src/switch.dart
@@ -47,11 +47,12 @@ extension Switch<T> on Stream<T> {
 extension SwitchLatest<T> on Stream<Stream<T>> {
   /// Emits values from the most recently emitted Stream.
   ///
-  /// When the source emits a stream the output will switch to emitting events
+  /// When the source emits a stream, the output will switch to emitting events
   /// from that stream.
   ///
-  /// If the source stream is a broadcast stream, the result stream will be as
-  /// well, regardless of the types of streams emitted.
+  /// Whether the source stream is a single-subscription stream or a
+  /// broadcast stream, the result stream will be the same kind of stream,
+  /// regardless of the types of streams emitted.
   Stream<T> switchLatest() {
     var controller = isBroadcast
         ? StreamController<T>.broadcast(sync: true)
@@ -75,8 +76,8 @@ extension SwitchLatest<T> on Stream<Stream<T>> {
         innerSubscription = subscription;
       }
 
-      final outerSubscription =
-          listen(null, onError: controller.addError, onDone: () {
+      var addError = controller.addError;
+      final outerSubscription = listen(null, onError: addError, onDone: () {
         outerStreamDone = true;
         if (innerSubscription == null) controller.close();
       });
@@ -84,12 +85,30 @@ extension SwitchLatest<T> on Stream<Stream<T>> {
         var currentSubscription = innerSubscription;
         if (currentSubscription != null) {
           innerSubscription = null;
-          outerSubscription.pause();
-          currentSubscription.cancel().whenComplete(() {
-            outerSubscription.resume();
-            listenToInnerStream(innerStream);
-          });
-          return;
+          try {
+            currentSubscription.cancel().catchError(addError).whenComplete(() {
+              if (!isBroadcast && !controller.hasListener) {
+                // Result single-subscription stream subscription was cancelled
+                // while waiting for previous innerStream cancel.
+                //
+                // Ensure that the last received stream is also listened to and
+                // cancelled, then do nothing further.
+                // TODO(lrn): When SDK 2.14 is available, use `.ignore()`.
+                innerStream
+                    .listen(null)
+                    .cancel()
+                    .then(_ignore, onError: _ignore);
+                return;
+              }
+              outerSubscription.resume();
+              listenToInnerStream(innerStream);
+            });
+            outerSubscription.pause();
+            return;
+          } catch (error, stack) {
+            // The cancel call threw synchronously.
+            controller.addError(error, stack);
+          }
         }
         listenToInnerStream(innerStream);
       });
@@ -105,16 +124,20 @@ extension SwitchLatest<T> on Stream<Stream<T>> {
           };
       }
       controller.onCancel = () {
+        var _innerSubscription = innerSubscription;
         var cancels = [
           if (!outerStreamDone) outerSubscription.cancel(),
-          if (innerSubscription != null) innerSubscription!.cancel(),
+          if (_innerSubscription != null) _innerSubscription.cancel(),
         ]
           // Handle opt-out nulls
           ..removeWhere((Object? f) => f == null);
         if (cancels.isEmpty) return null;
-        return Future.wait(cancels).then((_) => null);
+        return Future.wait(cancels).then(_ignore);
       };
     };
     return controller.stream;
   }
 }
+
+/// Helper function to ignore future callback
+void _ignore(_, [__]) {}