+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Wait for the future returned from `StreamSubscription.cancel()` before
   listening to the subsequent stream in `switchLatest` and `switchMap`.
-  
+
 ## 2.0.0
 
 - Migrate to null safety.